// src/app/Model/user.model.ts
export interface AppUser {
    id: bigint;
    /*nom: string;
    prenom: string;*/
    role: string;
    email: string;
    motDePasse: string;
    /*telephone: string;
    adresse: string;
    lieuDeResidence: string;
    photo:number; // Type pour représenter une photo (File)
    cv: number; // Type pour représenter un CV (File)
    lettreMotivation: string;*/
}